% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ab_arguments}
\alias{ab_arguments}
\title{Bayesian A/B Arguments}
\arguments{
\item{data_dists}{A named list of distribution objects. This list specifies the
distributions that are used to generate data in simulations.
Currently, the list must only have elements named \code{'a'}
and \code{'b'}. See Details for more information.}

\item{priors}{A named list of distribution objects. This list specifies the
distributions that are used as priors when estimating some parameter
from the data generating distribution. Currently, the list must only
have elements named \code{'a'} and \code{'b'}. See Details for more information.}

\item{loss_threshold}{A positive number that identifies a bound for the expected
loss for each variant. Once the expected loss is below this
bound, the experiment is concluded.}

\item{obs_per_round}{A positive number that represents how many observations, across both variants,
are generated before we update the prior distributions and
evaluate the expected loss. This number must be divisible by
the number of variants used. Default is \code{1000}.}

\item{max_rounds}{A positive integer that specifies the maximum number of
times that we will evaluate the expected loss on both experiments.
Default is \code{100}.}

\item{sim_batch_size}{A positive integer that specifies how much data is simulated
when evaluating the expected loss for variants that do
not have an analytic solution (i.e. normal data).}
}
\description{
Provide documentation for arguments commonly used in this package
}
