% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_priors.R
\name{approx_solver}
\alias{approx_solver}
\title{find_percentile}
\usage{
approx_solver(mu, bound, desired_support, cdf_fun, lower_sigma, upper_sigma,
  lower.tail = TRUE, tolerance = 0.01)
}
\arguments{
\item{mu}{The expected value of x (see details)}

\item{bound}{The desired maximum of the distribution}

\item{desired_support}{The amount of probability mass to be more extreme than \code{bound}}

\item{cdf_fun}{A CDF}

\item{lower_sigma}{A guess of the lower bound on the standard deviation}

\item{upper_sigma}{A guess of the upper bound on the standard deviation}

\item{lower.tail}{A logical that indicates whether to find the area of the upper or lower tail.
Default is \code{TRUE}.}

\item{tolerance}{Return \code{NULL}, if we cannot find a value that gives support
within \code{tolerance} of \code{desired_support}}
}
\value{
The optimal standard deviation
}
\description{
Determine the standard deviation of a distribution that will provide
             the desired amount of support at a certain bound.
}
