% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{b_gt_a}
\alias{b_gt_a}
\title{Probability Variant B is Greater Than Variant A}
\usage{
b_gt_a(dist_a, dist_b, theta_a, theta_b, exact = NULL, ...)
}
\arguments{
\item{dist_a}{Some distribution object (see examples)}

\item{dist_b}{Some distribution object (see examples)}

\item{theta_a}{A vector of simulated values from \code{dist_a}}

\item{theta_b}{A vector of simulated values from \code{dist_b}}

\item{exact}{A boolean that indicates whether the calculation should be approximated
using the normal distribution. Default is \code{NULL}, which means
that it will use the normal distribution if there is sufficient data.}

\item{...}{Arguments to be passed onto other methods}
}
\value{
A numeric value
}
\description{
Given two distributions, find the probability that the expected
             value of variant B is greater than the expected value of variant A
}
