% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dists.R
\name{bernoulli_dist}
\alias{bernoulli_dist}
\title{Bernoulli Distribution}
\usage{
bernoulli_dist(rate)
}
\arguments{
\item{rate}{The rate of the bernoulli distribution (between 0 and 1)}
}
\value{
A \code{bernoulli_dist} object
}
\description{
Create a bernoulli distribution object, used as the data generating
             distribution in tests where the metric is a probability of an
             event happening.
}
