% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_priors.R
\name{beta_cdf}
\alias{beta_cdf}
\title{CDF of Parameterized Beta Distribution}
\usage{
beta_cdf(mu, sigma, bound, lower.tail = TRUE)
}
\arguments{
\item{mu}{The expected value of x (see details)}

\item{sigma}{The standard deviation of x (see details)}

\item{bound}{The quantile of the distribution of interest}

\item{lower.tail}{A logical that indicates whether to find the area of the upper or lower tail.
Default is \code{TRUE}.}
}
\value{
The area under the desired tail.
}
\description{
Calculate the area to the left or right of \code{bound} for
             a beta distribution parameterized by the expected value and
             standard deviation.
}
\details{
If x ~ Beta(A, B), the \code{mu} = E[x] and \code{sigma^2} = Var(x)
}
