% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dists.R
\name{beta_dist}
\alias{beta_dist}
\title{Beta Distribution}
\usage{
beta_dist(alpha, beta)
}
\arguments{
\item{alpha}{The non-negative alpha parameter of the beta distribution}

\item{beta}{The non-negative beta parameter of the beta distribution}
}
\value{
An object of class \code{'beta'}.
}
\description{
Create a beta distribution object, used as the prior for metrics that
             have a bernoulli distribution.
}
