% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_priors.R
\name{calc_beta_dist}
\alias{calc_beta_dist}
\title{Calculate Parameters For Beta Distribution}
\usage{
calc_beta_dist(mu, sigma)
}
\arguments{
\item{mu}{The expected value of x (see details)}

\item{sigma}{The standard deviation of x (see details)}
}
\value{
An object of class \code{beta_dist}
}
\description{
Calculate the parameters for a beta distribution parameterized
             by the expected value and standard deviation.
}
\details{
If x ~ Beta(alpha, beta), then \code{mu} = E[x] and \code{sigma^2} = Var(x)
}
