% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_priors.R
\name{calc_gamma_dist}
\alias{calc_gamma_dist}
\title{Calculate Parameters For Gamma Distribution}
\usage{
calc_gamma_dist(mu, sigma)
}
\arguments{
\item{mu}{The expected value of x (see details)}

\item{sigma}{The standard deviation of x (see details)}
}
\value{
A named list
}
\description{
This function determines the parameters for a gamma distribution
             from the desired mean
}
\details{
If x ~ Gamma(alpha, beta), then \code{mu} = E[x] and \code{sigma^2} = Var(x)
}
