% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_empty_dt}
\alias{create_empty_dt}
\title{Create an Empty Data Table}
\usage{
create_empty_dt(num_rows, column_names, fill_value = -1)
}
\arguments{
\item{num_rows}{A positive integer that specifies the number of rows for the data table}

\item{column_names}{A character vector that specifies the number and names of the columns.}

\item{fill_value}{Default is -1. Specifies what value should fill the data table}
}
\value{
A data.table
}
\description{
Create an empty data table with a specified number of rows and columns.
             You can also choose the value that fills the data table.
}
