% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{expected_loss_b}
\alias{expected_loss_b}
\title{Expected Loss of Choosing Variant B}
\usage{
expected_loss_b(dist_a, dist_b, theta_a, theta_b, method = c("absolute",
  "percent"), ...)
}
\arguments{
\item{dist_a}{The distribution object for variant A}

\item{dist_b}{The distribution object for variant B}

\item{theta_a}{A vector of simulated values from \code{dist_a}}

\item{theta_b}{A vector of simulated values from \code{dist_b}}

\item{method}{One of \code{'absolute'} or \code{'percent'} that indicates
whether the loss function takes the absolute difference
or the percent difference between \code{theta_a} and \code{theta_b}}

\item{...}{Arguments to be passed onto other methods}
}
\value{
A numeric
}
\description{
This function calculates the expected loss of choosing variant B
}
