% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dists.R
\name{gamma_dist}
\alias{gamma_dist}
\title{Gamma Distribution}
\usage{
gamma_dist(alpha, beta)
}
\arguments{
\item{alpha}{The non-negative alpha (shape) parameter of the gamma distribution}

\item{beta}{The non-negative beta (rate) parameter of the gamma distribution}
}
\value{
A normal_unknown dist object
}
\description{
Create a gamma distribution object, used as the prior for
             metrics that have a poisson distribution.
}
