% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_actual_loss}
\alias{get_actual_loss}
\title{Actual Loss of an Experiment}
\usage{
get_actual_loss(data_dists, selected_variant)
}
\arguments{
\item{data_dists}{A named list of distribution objects. This list specifies the
distributions that are used to generate data in simulations.
Currently, the list must only have elements named \code{'a'}
and \code{'b'}. See Details for more information.}

\item{selected_variant}{A string that identifies the winning variant.}
}
\value{
Numeric
}
\description{
Once the experiment has concluded, measure the loss of the decision
}
