% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_dists.R
\name{get_data_dists}
\alias{get_data_dists}
\title{Sample Data Distributions}
\usage{
get_data_dists(dist, n, num_variants, ...)
}
\arguments{
\item{dist}{A prior distribution}

\item{n}{The number of distributions to sample}

\item{num_variants}{The number of variants to use}

\item{...}{Arguments to be used by other methods}
}
\value{
A list of data generating distributions for multiple variants
}
\description{
Given a prior distribution, sample many data generating distributions
}
