% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get Bayesian A/B Testing Metrics}
\usage{
get_metrics(posteriors, sim_batch_size, method = c("absolute", "percent"))
}
\arguments{
\item{posteriors}{A list of distribution objects}

\item{sim_batch_size}{How many observations of data to simulate}

\item{method}{What type of loss to calculate?}
}
\value{
A named list with the metrics
}
\description{
Calculate various metrics
}
