% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_priors.R
\name{get_supported_beta}
\alias{get_supported_beta}
\title{Get Beta Distribution With Desired Support}
\usage{
get_supported_beta(mu, bound, desired_support = 0.05)
}
\arguments{
\item{mu}{The expected value of x (see details)}

\item{bound}{The desired maximum of the distribution}

\item{desired_support}{The amount of probability mass to be more extreme than \code{bound}}
}
\value{
An object of class \code{beta_dist}
}
\description{
Determines the parameters for a beta distribution given an expected value
             and desired support at some bound.
}
\details{
If x ~ Beta(alpha, beta), then \code{mu} = E[x] and \code{sigma^2} = Var(x)
}
