% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_priors.R
\name{get_supported_gamma}
\alias{get_supported_gamma}
\title{Get parameters of gamma distribution from expected and maximum rate}
\usage{
get_supported_gamma(mu, bound, desired_support = 0.05, lower.tail = FALSE)
}
\arguments{
\item{mu}{The expected value of x (see details)}

\item{bound}{The desired maximum of the distribution}

\item{desired_support}{The amount of probability mass to be more extreme than \code{bound}}

\item{lower.tail}{A boolean that indicates whether we want to get the support on the upper or lower tail}
}
\value{
A \code{gamma_dist} object
}
\description{
This function determines the parameters for a gamma distribution
}
\details{
If x ~ Gamma(alpha, beta), then \code{mu} = E[x] and \code{sigma^2} = Var(x)
}
