% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_priors.R
\name{get_supported_normal_gamma}
\alias{get_supported_normal_gamma}
\title{Get Normal Gamma Distribution With Desired Support}
\usage{
get_supported_normal_gamma(mu, bound_mu, sigma, bound_sigma,
  desired_support = 0.05)
}
\arguments{
\item{mu}{The expected value of x (see details)}

\item{bound_mu}{An upper bound on the likely value of x (see details)}

\item{sigma}{The standard deviation of x (see details)}

\item{bound_sigma}{The upper bound on the likely value of T (see details)}

\item{desired_support}{The amount of probability mass to be more extreme than \code{bound}}
}
\value{
An object of class \code{normal_gamma_dist}
}
\description{
Determines the parameters for a normal gamma distribution given the expected values
             of the distribution and bounds on their maximum values
}
\details{
If (x, T) ~ NormalGamma(mu, lambda, alpha, beta), then \code{mu} = E[x],
         \code{tau} = E[T], \code{sigma_mu^2} = Var(x), and \code{sigma_tau^2} = Var(T)
}
