% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study.R
\name{investigate_simulations}
\alias{investigate_simulations}
\title{Investigate Properties of Bayesian A/B Testing}
\usage{
investigate_simulations(num_sims, priors, loss_threshold, data_dists = NULL,
  sampling_distribution = NULL, obs_per_round = 1000, max_rounds = 100,
  sim_batch_size = 1e+05, num_cores = NULL)
}
\arguments{
\item{num_sims}{A positive integer that specifies how many simulations to perform.}

\item{priors}{A named list of distribution objects. This list specifies the
distributions that are used as priors when estimating some parameter
from the data generating distribution. Currently, the list must only
have elements named \code{'a'} and \code{'b'}. See Details for more information.}

\item{loss_threshold}{A positive number that identifies a bound for the expected
loss for each variant. Once the expected loss is below this
bound, the experiment is concluded.}

\item{data_dists}{A named list of distribution objects. This list specifies the
distributions that are used to generate data in simulations.
Currently, the list must only have elements named \code{'a'}
and \code{'b'}. See Details for more information.}

\item{sampling_distribution}{An list of distribution objects that specifies how the data generating
distributions should be created.}

\item{obs_per_round}{A positive number that represents how many observations, across both variants,
are generated before we update the prior distributions and
evaluate the expected loss. This number must be divisible by
the number of variants used. Default is \code{1000}.}

\item{max_rounds}{A positive integer that specifies the maximum number of
times that we will evaluate the expected loss on both experiments.
Default is \code{100}.}

\item{sim_batch_size}{A positive integer that specifies how much data is simulated
when evaluating the expected loss for variants that do
not have an analytic solution (i.e. normal data).}

\item{num_cores}{How many cores to use in the parallelization of tests. Default is \code{NULL}, which
means no parallelization.}
}
\value{
A list containing two data.tables: one with summary statistics for each simulation
        and one with the averages over all of the simulations.
}
\description{
Run multiple simulations of an A/B test in order to evaluate various properties
}
