% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dists.R
\name{normal_dist}
\alias{normal_dist}
\title{Normal Distribution}
\usage{
normal_dist(mu, sigma)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The non-negative standard deviation of the normal distribution}
}
\value{
A \code{normal_dist} object
}
\description{
Create a normal distribution object, used as the data generating
             distribution in tests where the metric is a continuous value.
}
