% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dists.R
\name{normal_gamma_dist}
\alias{normal_gamma_dist}
\title{Normal-Gamma Distribution}
\usage{
normal_gamma_dist(mu, lambda, alpha, beta)
}
\arguments{
\item{mu}{The mean of the normal-gamma distribution}

\item{lambda}{The non-negative lambda parameter of the normal-gamma distribution}

\item{alpha}{The non-negative alpha parameter of the normal-gamma distribution}

\item{beta}{The non-negative beta parameter of the normal-gamma distribution}
}
\value{
A normal_unknown dist object
}
\description{
Create a normal_gamma distribution object, used as the prior for
             metrics that have a normal distribution.
}
