% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot Beta Distributions}
\usage{
plot_beta(betas, title = "Beta Distribution",
  xlab = "Rate that the Event Occurs", ylab = "Density of that Rate",
  color = "#f65335", level = 0.99)
}
\arguments{
\item{betas}{A list of lists of beta distributions}

\item{title}{The title of the plot}

\item{xlab}{The title of the x axis}

\item{ylab}{The title of the y axis}

\item{color}{The color for the plot}

\item{level}{The desired amount of area between the lower and upper bounds. Default is \code{0.99}.}
}
\value{
NULL. A plot is generated
}
\description{
This function plots the densities of multiple beta distributions
}
