% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot Gamma Distributions}
\usage{
plot_gamma(gammas, title = "Density of Gamma Distribution",
  color = "#f65335", support_level = 0.99)
}
\arguments{
\item{gammas}{A list of lists of gamma distributions}

\item{title}{The title of the plot}

\item{color}{The color for the plot}

\item{support_level}{The desired amount of area between the lower and upper bounds. Default is \code{0.99}.}
}
\value{
NULL. A plot is generated
}
\description{
This function plots the densities of multiple gamma distributions
}
