% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_normal}
\alias{plot_normal}
\title{Plot Normal Distributions}
\usage{
plot_normal(normals)
}
\arguments{
\item{normals}{A list of lists of normal distributions}
}
\value{
NULL. A plot is generated
}
\description{
This function allows you to visualize the densities of a normal distribution
}
