% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_relative_gain}
\alias{plot_relative_gain}
\title{Plot Relative Gain}
\usage{
plot_relative_gain(dists, sim_batch_size = 1e+05,
  title = "Cumulative Density of B / A")
}
\arguments{
\item{dists}{A list of distribution objects, with elements named \code{'a'} and \code{'b'}}

\item{sim_batch_size}{The number of objects to simulate}
}
\value{
A plot
}
\description{
Plot the cumulative density of the ratio of the metric under variant B to the metric under variant A
}
