% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dists.R
\name{poisson_dist}
\alias{poisson_dist}
\title{Poisson Distribution}
\usage{
poisson_dist(rate)
}
\arguments{
\item{rate}{The non-negative mean of the poisson distribution}
}
\value{
A \code{poisson_dist} object
}
\description{
Create a poisson distribution object, used as the data generating
             distribution in tests where the metric is a count of an event happening.
}
