% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sim_effect_size}
\alias{sim_effect_size}
\title{Simulate the effect size of variant B - variant A}
\usage{
sim_effect_size(theta_a, theta_b)
}
\arguments{
\item{theta_a}{A vector of draws from the posterior of a}

\item{theta_b}{A vector of draws from the posterior of b}
}
\value{
A list containing the 2.5%, mean, and 97.5% of the effect size
}
\description{
This function simulates the effect size (B - A)
}
