% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{simulate_ab_test}
\alias{simulate_ab_test}
\title{Simulate a Bayesian A/B Test}
\usage{
simulate_ab_test(data_dists, priors, loss_threshold, obs_per_round = 1000,
  max_rounds = 100, sim_batch_size = 1e+05)
}
\arguments{
\item{data_dists}{A named list of distribution objects. This list specifies the
distributions that are used to generate data in simulations.
Currently, the list must only have elements named \code{'a'}
and \code{'b'}. See Details for more information.}

\item{priors}{A named list of distribution objects. This list specifies the
distributions that are used as priors when estimating some parameter
from the data generating distribution. Currently, the list must only
have elements named \code{'a'} and \code{'b'}. See Details for more information.}

\item{loss_threshold}{A positive number that identifies a bound for the expected
loss for each variant. Once the expected loss is below this
bound, the experiment is concluded.}

\item{obs_per_round}{A positive number that represents how many observations, across both variants,
are generated before we update the prior distributions and
evaluate the expected loss. This number must be divisible by
the number of variants used. Default is \code{1000}.}

\item{max_rounds}{A positive integer that specifies the maximum number of
times that we will evaluate the expected loss on both experiments.
Default is \code{100}.}

\item{sim_batch_size}{A positive integer that specifies how much data is simulated
when evaluating the expected loss for variants that do
not have an analytic solution (i.e. normal data).}
}
\value{
A list that contains the name of the winning variant, the number of observations used,
        the loss of the decision, whether the test finished, the metrics from each round, and
        the raw data.
}
\description{
Given true data generating distributions and prior distributions for
             variants A and B, simulate the data, calculate the necessary statistics
             and declare one of the tests a winner.
}
\details{
In order to create \code{data_dists} and \code{priors}, you need to use the following
         for data generating distributions: \code{\link{bernoulli_dist}},
         \code{\link{normal_dist}}, \code{\link{poisson_dist}} and the following for prior
         distributions \code{\link{beta_dist}}, \code{\link{normal_gamma_dist}},
         \code{\link{beta_dist}},
}
\seealso{
\code{\link{bernoulli_dist}} \code{\link{normal_dist}} \code{\link{poisson_dist}}
         \code{\link{beta_dist}} \code{\link{normal_gamma_dist}} \code{\link{beta_dist}}
}
