% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_moments.R, R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\alias{compute_moments}
\alias{simulate_data}
\title{Simulate Data According to Some Distribution}
\usage{
compute_moments(dist)

simulate_data(dist, n, ...)
}
\arguments{
\item{dist}{An object of class \code{'beta_dist'}, \code{'normal_gamma_dist'},
\code{'gamma_dist'} that specifies the parameters of some distribution}

\item{n}{How many data points to simulate}

\item{...}{Arguments to be passed onto other methods}

\item{dist}{An object of class \code{'beta_dist'}, \code{'normal_gamma_dist'},
\code{'gamma_dist'}, \code{'bernouilli_dist'}, \code{'normal_dist'},
or \code{'poisson_dist'} that specifies the parameters of some distribution}
}
\value{
A list

A vector of simulated data.
}
\description{
Simulate a vector of data from a given distribution object.

Simulate a vector of data from a given distribution object.
}
