% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping_functions.R
\name{single_loss_stop}
\alias{single_loss_stop}
\title{Decide whether to stop the experiment using the last loss values}
\usage{
single_loss_stop(losses, loss_threshold)
}
\arguments{
\item{losses}{A named vector of the losses for each variant.}

\item{loss_threshold}{Choose a variant once the loss goes beneath this value.}
}
\value{
A list containg two named elements: \code{'stop'} (a boolean) and
        \code{'winner'}, which is \code{NULL} if \code{'stop'} is \code{FALSE}
        else it is the name of the winning variant
}
\description{
This function decides whether or not to stop the experiment and
             declare one of the variants as the winner.
}
