% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_priors.R
\name{update_prior}
\alias{update_prior}
\title{Update Prior Parameters}
\usage{
update_prior(prior, evidence = NULL, stats = NULL, ...)
}
\arguments{
\item{prior}{An object of class \code{'beta_dist'}, \code{'normal_gamma_dist'}
, or \code{'gamma_dist'} that specifies the parameters of some distribution}

\item{evidence}{A numeric vector that contains observed data. Default is \code{NULL},
will override \code{stats} if specified.}

\item{stats}{An object of class \code{'beta_stats'}, \code{'normal_gamma_stats'}, or
\code{'gamma_stats'} that contains sufficient statistics for the
update. Default is \code{NULL}, will be ignored if \code{evidence}
is specified.}

\item{...}{Arguments to be passed onto other methods}
}
\description{
Use observed data to update parameters of prior distribution
}
