% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_priors.R
\name{update_priors}
\alias{update_priors}
\title{Update the hyper-parameters of prior distributions}
\usage{
update_priors(priors, evidence_dt = NULL, stats_dt = NULL)
}
\arguments{
\item{priors}{A named list of distribution objects. This list specifies the
distributions that are used as priors when estimating some parameter
from the data generating distribution. Currently, the list must only
have elements named \code{'a'} and \code{'b'}. See Details for more information.}

\item{evidence_dt}{A data.table containing the raw data generated by each variant.
Column names must be identical to variant names in \code{priors}}

\item{stats_dt}{A data.table containing statistics about the raw data generated by each
variant.}
}
\value{
A list of distribution objects that represented the updated prior distributions
}
\description{
This function updates the hyper-parameters of our prior distributions
}
