% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_priors.R
\name{validate_dt}
\alias{validate_dt}
\title{Check That Data Table Matches Expectations}
\usage{
validate_dt(dt, expected_cols)
}
\arguments{
\item{dt}{A data.table}

\item{expected_cols}{A vector of column names that \code{dt} must have}
}
\value{
\code{NULL} if successful, else a fatal error
}
\description{
Use this function to assert certain properties of a data.table
}
