% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate Inputs to Bayesian A/B Simulation}
\usage{
validate_inputs(data_dists, priors, obs_per_round)
}
\arguments{
\item{data_dists}{A named list of distribution objects. This list specifies the
distributions that are used to generate data in simulations.
Currently, the list must only have elements named \code{'a'}
and \code{'b'}. See Details for more information.}

\item{priors}{A named list of distribution objects. This list specifies the
distributions that are used as priors when estimating some parameter
from the data generating distribution. Currently, the list must only
have elements named \code{'a'} and \code{'b'}. See Details for more information.}

\item{obs_per_round}{A positive number that represents how many observations, across both variants,
are generated before we update the prior distributions and
evaluate the expected loss. This number must be divisible by
the number of variants used. Default is \code{1000}.}
}
\value{
NULL if all of the tests pass. Else, it will fail loudly.
}
\description{
Check the validity of the parameters being used in the simulation.
}
